requirements.txt
# Core Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Machine Learning
scikit-learn>=1.3.0

# Database
mysql-connector-python>=8.0.33

# Configuration
pyyaml>=6.0

# Visualization
matplotlib>=3.7.0
seaborn>=0.12.0

# Utilities
python-dateutil>=2.8.2
joblib>=1.3.0

# CLI
argparse>=1.4.0

# Jupyter (optional, for notebooks)
jupyter>=1.0.0
ipykernel>=6.25.0

# Progress bars (nice to have)
tqdm>=4.65.0

